Coerce to year month

Date, POSIXct, and POSIXlt are converted directly by extracting the year and month from x. Any day, hour, minute, or second components are dropped. Time zone information is not retained.

Integer and double input are assumed to be the number of months since the Unix origin of 1970-01-01.

Character input is assumed to be provided in a format containing only information about the year and month, such as "1970-01" or "Jan 1970". It is parsed using the defaults of ym_parse().

as_ym(x, ...)

# S3 method for Date
as_ym(x, ...)

# S3 method for POSIXct
as_ym(x, ...)

# S3 method for POSIXlt
as_ym(x, ...)

# S3 method for integer
as_ym(x, ...)

# S3 method for double
as_ym(x, ...)

# S3 method for character
as_ym(x, ...)

Arguments

x
[vector]

An object to coerce to ym.

...
Not used.

Examples

# Extra information such as days, hours, or time zones are dropped
as_ym(as.Date("2019-05-03"))

#> <ym[1]>
#> [1] 2019-05

as_ym(as.POSIXct("2019-03-04 01:01:01", tz = "America/New_York"))

#> <ym[1]>
#> [1] 2019-03

# Integers are interpreted as the number of months since 1970-01-01
as_ym(0L)

#> <ym[1]>
#> [1] 1970-01

as_ym(12L)

#> <ym[1]>
#> [1] 1971-01