Sequence generation

This seq() method generates regular sequences of ym objects. It functions the same as the default seq() method, except that fractional sequences that might be generated by seq(to =, length.out =) are not allowed.

seq() methods require a from argument, and any combination of two of the other arguments. The only combination not allowed is length.out and along.with, because they represent the same information.

# S3 method for ym
seq(from, to, by, length.out, along.with, ...)

Arguments

from
[ym]

A ym.

to
[ym]

Optional. A ym.

by
[integer(1)]

Optional. A single integer representing the step size. Must be negative if to is supplied and the sequence between from and to is decreasing.

length.out
[integer(1)]

Optional. A single integer representing the final output size. If to is supplied, this must divide the distance between from and to into equally spaced pieces, otherwise an error is thrown.

along.with
[vector]

Optional. A vector to take the length of. The length of the vector is used as the length.out value.

...
Not used.

Examples

x <- ym(1970, 1)

seq(x, by = 3, length.out = 10)

#> <ym[10]>
#>  [1] 1970-01 1970-04 1970-07 1970-10 1971-01 1971-04 1971-07 1971-10 1972-01
#> [10] 1972-04

seq(x, to = ym(1980, 1), by = 6)

#> <ym[21]>
#>  [1] 1970-01 1970-07 1971-01 1971-07 1972-01 1972-07 1973-01 1973-07 1974-01
#> [10] 1974-07 1975-01 1975-07 1976-01 1976-07 1977-01 1977-07 1978-01 1978-07
#> [19] 1979-01 1979-07 1980-01

seq(x, to = ym(1980, 1), length.out = 5)

#> <ym[5]>
#> [1] 1970-01 1972-07 1975-01 1977-07 1980-01

# 4 values between 1970-01 and 1970-04, there
# is no way to divide these evenly into 3 equally spaced pieces!
try(seq(x, to = ym(1970, 4), length.out = 3))

#> Error : `length.out` / `along.with` must generate a non-fractional sequence between `from` and `to`.