Parse a character vector as year month

ym_parse() is a parser for turning character input into ym objects. The character method for as_ym() is very strict, and only parses input of the form "YYYY-MM", erroring on any failures. ym_parse() is more flexible and will only issue a warning if parsing fails.

Internally:

"-%d" is appended to the format, and "-01" is appended to x.

An attempt to parse as a Date is then made.

The resulting Date is converted to ym, with a warning if any input failed to parse. Failures result in NA.

ym_parse(x, format = "%Y-%m")

Arguments

x
[character]

A character vector to coerce to ym.

format
[character(1)]

A format to parse character input with. Should generally only consist of format tokens related to year or month. Common formats are "%Y-%m", "%b %Y", and "%Y %b".

Examples

ym_parse("1970-01")

#> <ym[1]>
#> [1] 1970-01

ym_parse("1970 Jan", format = "%Y %b")

#> <ym[1]>
#> [1] 1970-01

# Unparseable input results in `NA`, with a warning
try(ym_parse(c("1970-00", "1970-01")))

#> Warning: Unable to parse to year month at location 1. Returning `NA` at that location.

#> <ym[2]>
#> [1] NA      1970-01